Non-trivial default constructor initializing val=dot=0
The advantage of a trivial default constructor is that the type
can be used in unions. However, unions are seen very rarely and
if they are used to perform bit-tricks, it's actually good to
alert the user.

On the other hand, a non-trivial default constructor allows to
zero the value and dot value.